Modify the code (most importantly remove delegating constructors) so that the code compiles on g++-4.6.
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/branches/sgopt@3583 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/opt/optimizer/numrecipes/Frprmn.hpp
@@ -67,7 +67,7 @@ struct Frprmn : Linemethod<T>
         
         func.df(p, xi);
         
-        for (int j = 0; j < n; j++)
+        for (size_t j = 0; j < n; j++)
         {
             g[j] = -xi[j];
             xi[j] = g[j];
